client/requirements.txt
colorzero==2.0
gpiozero==2.0
lgpio==0.0.0.2
pyserial==3.5
python-dotenv==1.0.0
RPi.GPIO==0.7.1
psutil==5.9.8
dbus-python==1.3.2
pycairo==1.26.0
pygobject==3.48.2
PyBluez==0.23
pycryptodome==3.20.0
